move task builder to index, move task action to named task file
